backend/requirements.txt
flask
matplotlib
numpy
pandas
torch
torchvision
Pillow
opencv-python
opencv-python-headless
gunicorn